requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
azure-common==1.1.28
azure-core==1.30.1
azure-mgmt-compute==30.6.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.19.1
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.5
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
comm==0.1.4
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
Django==5.0.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.9.0
Flask==2.2.3
Flask-Mail==0.9.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
image==1.5.33
install==1.3.5
ipykernel==6.25.1
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mimerender==0.6.0
mistune==3.0.1
MouseInfo==0.1.3
multidict==6.0.5
nbclient==0.7.4
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.26.1
openai==0.28.1
opencv-python==4.9.0.80
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==2.6.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-mimeparse==1.6.0
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robotframework==7.0
robotframework-pythonlibcore==4.3.0
robotframework-seleniumlibrary==6.2.0
rpds-py==0.10.0
rubicon-objc==0.4.7
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.2
selenium==4.18.1
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post10
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.9.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.0.4
virtualenv==20.17.1
voila==0.5.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.8
wsproto==1.2.0
yarl==1.9.4
